Run product actions concurrently.
